+++ .changeset/where-compiler-extraction.md
@@ -1,0 +1,5 @@
+---
+"@danceroutine/tango-orm": patch
+---
+
+Existing `QuerySet` filter, exclude, relation hydration, and prefetch SQL outputs remain unchanged while predicate, hydration-plan, and prefetch rendering move into internal compiler components.
